@@ -23,10 +23,11 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
-        "typo3/cms-core": "~7.6.0 || ~8.7.0",
-        "rkw/rkw-basics": "~8.7.53",
-        "rkw/rkw-ajax": "~8.7.6"
+        "php": ">=7.4",
+        "typo3/cms-core": "~9.5.0",
+		"madj2k/t3-core-extended": "~9.5.4",
+		"madj2k/t3-accelerator": "~9.5.2",
+		"madj2k/t3-ajax-api": "~9.5.0"
     },
     "autoload": {
         "psr-4": {
