@@ -14,254 +14,96 @@ namespace RKW\RkwAuthors\ViewHelpers;
  * The TYPO3 project - inspiring people to share!
  */
 
+use RKW\RkwAuthors\Domain\Model\Authors;
 use TYPO3\CMS\Fluid\Core\ViewHelper\AbstractViewHelper;
 use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
+use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithContentArgumentAndRenderStatic;
 
+/**
+ * Class GetCombinedNameViewHelper
+ *
+ * @author Steffen Kroggel <developer@steffenkroggel.de>
+ * @copyright RKW Kompetenzzentrum
+ * @package RKW_RkwAuthors
+ * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
+ * @todo write tests
+ */
+class GetCombinedNameViewHelper extends AbstractViewHelper
+{
 
-$currentVersion = \TYPO3\CMS\Core\Utility\VersionNumberUtility::convertVersionNumberToInteger(TYPO3_version);
-if ($currentVersion < 8000000) {
+    use CompileWithContentArgumentAndRenderStatic;
 
-    if (\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::isLoaded('rkw_search')) {
-        /**
-         * Class GetCombinedNameViewHelper
-         *
-         * @author Steffen Kroggel <developer@steffenkroggel.de>
-         * @copyright Rkw Kompetenzzentrum
-         * @package RKW_RkwAuthors
-         * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
-         * @deprecated
-         */
-        class GetCombinedNameViewHelper extends AbstractViewHelper
-        {
+    /**
+     * Output is escaped already. We must not escape children, to avoid double encoding.
+     *
+     * @var bool
+     */
+    protected $escapeChildren = false;
 
-            /**
-             * Build the full name
-             *
-             * @param \RKW\RkwSearch\OrientDb\Domain\Model\DocumentAuthors $author
-             * @param boolean $addTitleBefore
-             * @param boolean $addTitleAfter
-             * @return string
-             */
-            public function render($author, $addTitleBefore = true, $addTitleAfter = false)
-            {
-
-                return static::renderStatic(
-                    array(
-                        'author'         => $author,
-                        'addTitleBefore' => $addTitleBefore,
-                        'addTitleAfter'  => $addTitleAfter,
-                    ),
-                    $this->buildRenderChildrenClosure(),
-                    $this->renderingContext
-                );
-            }
-
-
-            /**
-             * Static rendering
-             *
-             * @param array $arguments
-             * @param \Closure $renderChildrenClosure
-             * @param RenderingContextInterface $renderingContext
-             * @return string
-             */
-            static public function renderStatic(array $arguments, \Closure $renderChildrenClosure, \TYPO3\CMS\Fluid\Core\Rendering\RenderingContextInterface $renderingContext)
-            {
-                $author = $arguments['author'];
-                $addTitleBefore = $arguments['addTitleBefore'];
-                $addTitleAfter = $arguments['addTitleAfter '];
-
-                $name = '';
-                if (
-                    ($author instanceof \RKW\RkwAuthors\Domain\Model\Authors)
-                    || ($author instanceof \RKW\RkwSearch\OrientDb\Domain\Model\DocumentAuthors)
-                ) {
-                    $name = $author->getLastName() ? $author->getLastName() : $author->getLastname();
-
-                    if ($author->getMiddleName() || $author->getMiddleName()) {
-                        $name = ($author->getMiddleName() ? $author->getMiddleName() : $author->getMiddlename()) . ' ' . $name;
-                    }
-
-                    if ($author->getFirstName() || $author->getFirstname()) {
-                        $name = ($author->getFirstName() ? $author->getFirstName() : $author->getFirstname()) . ' ' . $name;
-                    }
-
-                    if ($addTitleBefore) {
-                        if ($author->getTitleBefore()) {
-                            $name = $author->getTitleBefore() . ' ' . $name;
-                        }
-                    }
-
-                    if ($addTitleAfter) {
-                        if ($author->getTitleAfter()) {
-                            $name .= ', ' . $author->getTitleAfter();
-                        }
-                    }
-                }
-
-                return $name;
-            }
-        }
-
-    } else {
-
-        /**
-         * Class GetCombinedNameViewHelper
-         *
-         * @author Steffen Kroggel <developer@steffenkroggel.de>
-         * @copyright Rkw Kompetenzzentrum
-         * @package RKW_RkwAuthors
-         * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
-         * @deprecated
-         */
-        class GetCombinedNameViewHelper extends AbstractViewHelper
-        {
-
-            /**
-             * Build the full name
-             *
-             * @param \RKW\RkwAuthors\Domain\Model\Authors $author
-             * @param boolean $addTitleBefore
-             * @param boolean $addTitleAfter
-             * @return string
-             */
-            public function render($author, $addTitleBefore = true, $addTitleAfter = false)
-            {
-
-                return static::renderStatic(
-                    array(
-                        'author'         => $author,
-                        'addTitleBefore' => $addTitleBefore,
-                        'addTitleAfter'  => $addTitleAfter,
-                    ),
-                    $this->buildRenderChildrenClosure(),
-                    $this->renderingContext
-                );
-            }
-
-
-            /**
-             * Static rendering
-             *
-             * @param array $arguments
-             * @param \Closure $renderChildrenClosure
-             * @param RenderingContextInterface $renderingContext
-             * @return string
-             */
-            static public function renderStatic(array $arguments, \Closure $renderChildrenClosure, \TYPO3\CMS\Fluid\Core\Rendering\RenderingContextInterface $renderingContext)
-            {
-                $author = $arguments['author'];
-                $addTitleBefore = $arguments['addTitleBefore'];
-                $addTitleAfter = $arguments['addTitleAfter '];
-
-                $name = '';
-                if ($author instanceof \RKW\RkwAuthors\Domain\Model\Authors) {
-
-                    $name = $author->getLastName() ? $author->getLastName() : $author->getLastname();
-                    if ($author->getMiddleName() || $author->getMiddleName()) {
-                        $name = ($author->getMiddleName() ? $author->getMiddleName() : $author->getMiddlename()) . ' ' . $name;
-                    }
-
-                    if ($author->getFirstName() || $author->getFirstname()) {
-                        $name = ($author->getFirstName() ? $author->getFirstName() : $author->getFirstname()) . ' ' . $name;
-                    }
-
-                    if ($addTitleBefore) {
-                        if ($author->getTitleBefore()) {
-                            $name = $author->getTitleBefore() . ' ' . $name;
-                        }
-                    }
-
-                    if ($addTitleAfter) {
-                        if ($author->getTitleAfter()) {
-                            $name .= ', ' . $author->getTitleAfter();
-                        }
-                    }
-                }
-
-                return $name;
-            }
-        }
+    /**
+     * Initialize arguments.
+     *
+     * @return void
+     * @throws \TYPO3Fluid\Fluid\Core\ViewHelper\Exception
+     */
+    public function initializeArguments(): void
+    {
+        parent::initializeArguments();
+        $this->registerArgument('author', Authors::class, 'The author-object', true);
+        $this->registerArgument('addTitleBefore', 'bool', 'Adds the titles that belong before the name', false, false);
+        $this->registerArgument('addTitleAfter', 'bool', 'Adds the titles that belong after the name', false, false);
     }
 
 
-} else {
-
     /**
-     * Class GetCombinedNameViewHelper
+     * Static rendering
      *
-     * @author Steffen Kroggel <developer@steffenkroggel.de>
-     * @copyright Rkw Kompetenzzentrum
-     * @package RKW_RkwAuthors
-     * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
+     * @param array $arguments
+     * @param \Closure $renderChildrenClosure
+     * @param RenderingContextInterface $renderingContext
+     * @return string
      */
-    class GetCombinedNameViewHelper extends AbstractViewHelper
-    {
+    static public function renderStatic(
+        array $arguments,
+        \Closure $renderChildrenClosure,
+        RenderingContextInterface $renderingContext
+    ): string {
 
-        /**
-         * Build the full name
-         *
-         * @param \RKW\RkwAuthors\Domain\Model\Authors $author
-         * @param boolean $addTitleBefore
-         * @param boolean $addTitleAfter
-         * @return string
-         */
-        public function render($author, $addTitleBefore = true, $addTitleAfter = false)
-        {
+        /** @var \RKW\RkwAuthors\Domain\Model\Authors $author */
+        $author = $arguments['author'];
 
-            return static::renderStatic(
-                array(
-                    'author'         => $author,
-                    'addTitleBefore' => $addTitleBefore,
-                    'addTitleAfter'  => $addTitleAfter,
-                ),
-                $this->buildRenderChildrenClosure(),
-                $this->renderingContext
-            );
-        }
+        /** @var bool $addTitleBefore */
+        $addTitleBefore = $arguments['addTitleBefore'];
 
+        /** @var bool $addTitleAfter */
+        $addTitleAfter = $arguments['addTitleAfter'];
 
-        /**
-         * Static rendering
-         *
-         * @param array $arguments
-         * @param \Closure $renderChildrenClosure
-         * @param RenderingContextInterface $renderingContext
-         * @return string
-         */
-        static public function renderStatic(array $arguments, \Closure $renderChildrenClosure, RenderingContextInterface $renderingContext)
-        {
-            $author = $arguments['author'];
-            $addTitleBefore = $arguments['addTitleBefore'];
-            $addTitleAfter = $arguments['addTitleAfter '];
+        $name = '';
+        if ($author instanceof \RKW\RkwAuthors\Domain\Model\Authors) {
+            $name = $author->getLastName() ?: $author->getLastname();
 
-            $name = '';
-            if ($author instanceof \RKW\RkwAuthors\Domain\Model\Authors) {
-                $name = $author->getLastName() ? $author->getLastName() : $author->getLastname();
+            if ($author->getMiddleName()) {
+                $name = $author->getMiddlename() . ' ' . $name;
+            }
 
-                if ($author->getMiddleName() || $author->getMiddleName()) {
-                    $name = ($author->getMiddleName() ? $author->getMiddleName() : $author->getMiddlename()) . ' ' . $name;
-                }
+            if ($author->getFirstName()) {
+                $name = $author->getFirstName() . ' ' . $name;
+            }
 
-                if ($author->getFirstName() || $author->getFirstname()) {
-                    $name = ($author->getFirstName() ? $author->getFirstName() : $author->getFirstname()) . ' ' . $name;
-                }
-
-                if ($addTitleBefore) {
-                    if ($author->getTitleBefore()) {
-                        $name = $author->getTitleBefore() . ' ' . $name;
-                    }
-                }
-
-                if ($addTitleAfter) {
-                    if ($author->getTitleAfter()) {
-                        $name .= ', ' . $author->getTitleAfter();
-                    }
+            if ($addTitleBefore) {
+                if ($author->getTitleBefore()) {
+                    $name = $author->getTitleBefore() . ' ' . $name;
                 }
             }
 
-            return $name;
+            if ($addTitleAfter) {
+                if ($author->getTitleAfter()) {
+                    $name .= ', ' . $author->getTitleAfter();
+                }
+            }
         }
+
+        return $name;
     }
 }
 
