@@ -2,10 +2,6 @@
 
 namespace RKW\RkwAuthors\ViewHelpers;
 
-use \RKW\RkwBasics\Helper\Common;
-use \TYPO3\CMS\Extbase\Configuration\ConfigurationManagerInterface;
-use TYPO3\CMS\Extbase\Utility\DebuggerUtility;
-
 /*
  * This file is part of the TYPO3 CMS project.
  *
@@ -19,24 +15,59 @@ use TYPO3\CMS\Extbase\Utility\DebuggerUtility;
  * The TYPO3 project - inspiring people to share!
  */
 
+use RKW\RkwAuthors\Domain\Model\Authors;
+use Madj2k\CoreExtended\Utility\GeneralUtility;
+use TYPO3\CMS\Core\Log\Logger;
+use TYPO3\CMS\Core\Log\LogManager;
+use \TYPO3\CMS\Extbase\Configuration\ConfigurationManagerInterface;
+use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
+use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithContentArgumentAndRenderStatic;
+
 /**
  * Class IfContactFormIsActiveViewHelper
  *
- * @author Maximilian Fäßler <faesslerweb@web.de>
- * @copyright Rkw Kompetenzzentrum
+ * @author Maximilian Fäßler <maximilian@faesslerweb.de>
+ * @copyright RKW Kompetenzzentrum
  * @package RKW_RkwAuthors
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
+ * @todo write tests
  */
 class IfContactFormIsActiveViewHelper extends \TYPO3\CMS\Fluid\Core\ViewHelper\AbstractViewHelper
 {
 
+    use CompileWithContentArgumentAndRenderStatic;
+
     /**
-     * @param \RKW\RkwAuthors\Domain\Model\Authors $author
-     * @return bool
+     * Initialize arguments.
+     *
+     * @return void
+     * @throws \TYPO3Fluid\Fluid\Core\ViewHelper\Exception
      */
-    public function render(\RKW\RkwAuthors\Domain\Model\Authors $author)
+    public function initializeArguments(): void
     {
-        $settings = $this->getSettings();
+        parent::initializeArguments();
+        $this->registerArgument('author', Authors::class, 'The author-object', true);
+    }
+
+
+    /**
+     * Static rendering
+     *
+     * @param array $arguments
+     * @param \Closure $renderChildrenClosure
+     * @param RenderingContextInterface $renderingContext
+     * @return bool
+     * @throws \TYPO3\CMS\Extbase\Configuration\Exception\InvalidConfigurationTypeException
+     */
+    static public function renderStatic(
+        array $arguments,
+        \Closure $renderChildrenClosure,
+        RenderingContextInterface $renderingContext
+    ): bool {
+
+        /** @var \RKW\RkwAuthors\Domain\Model\Authors $author */
+        $author = $arguments['author'];
+        $settings = self::getSettings();
 
         //  if pageUid for contactForm is set
         if ($settings['contactForm']['pageUid']) {
@@ -45,9 +76,11 @@ class IfContactFormIsActiveViewHelper extends \TYPO3\CMS\Fluid\Core\ViewHelper\A
             if (
                 !\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::isLoaded('rkw_mailer')
             ) {
-                $this->getLogger()->log(\TYPO3\CMS\Core\Log\LogLevel::WARNING, sprintf('The optional contact form function of the rkw_authors extension is only available, if also the additional extension rkw_mailer are activated.'));
+                self::getLogger()->log(
+                    \TYPO3\CMS\Core\Log\LogLevel::WARNING,
+                    'The optional contact form function of the rkw_authors extension is only available, if also the additional extension rkw_mailer are activated.'
+                );
                 return false;
-                //===
             }
 
             // Error: If we have no authors email address AND no fallback
@@ -55,20 +88,24 @@ class IfContactFormIsActiveViewHelper extends \TYPO3\CMS\Fluid\Core\ViewHelper\A
                 !$author->getEmail()
                 && !$settings['contactForm']['mail']['fallback']['address']
             ) {
-                $this->getLogger()->log(\TYPO3\CMS\Core\Log\LogLevel::ERROR, sprintf('The contact form is activated, but the author with UID %s has either an email address nor a fallback is defined.', $author->getUid()));
+                self::getLogger()->log(
+                    \TYPO3\CMS\Core\Log\LogLevel::ERROR,
+                    sprintf(
+                        'The contact form is activated, but the author with UID %s has either an email address nor a fallback is defined.',
+                        $author->getUid()
+                    )
+                );
                 return false;
-                //===
             }
 
             // if nothing happen: show link to contact form
             return true;
-            //===
         }
 
         // do not show link to contact form
         return false;
-        //===
     }
+
 
     /**
      * Returns TYPO3 settings
@@ -77,20 +114,19 @@ class IfContactFormIsActiveViewHelper extends \TYPO3\CMS\Fluid\Core\ViewHelper\A
      * @return array
      * @throws \TYPO3\CMS\Extbase\Configuration\Exception\InvalidConfigurationTypeException
      */
-    protected function getSettings($which = ConfigurationManagerInterface::CONFIGURATION_TYPE_SETTINGS)
+    protected static function getSettings(string $which = ConfigurationManagerInterface::CONFIGURATION_TYPE_SETTINGS): array
     {
-        return Common::getTyposcriptConfiguration('Rkwauthors', $which);
-        //===
+        return GeneralUtility::getTypoScriptConfiguration('Rkwauthors', $which);
     }
+
 
     /**
      * Returns logger instance
      *
      * @return \TYPO3\CMS\Core\Log\Logger
      */
-    protected function getLogger()
+    protected static function getLogger(): Logger
     {
-        return \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Core\\Log\\LogManager')->getLogger(__CLASS__);
-        //===
+        return \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(LogManager::class)->getLogger(__CLASS__);
     }
 }
