@@ -13,38 +13,64 @@ namespace RKW\RkwAuthors\ViewHelpers;
  *
  * The TYPO3 project - inspiring people to share!
  */
-use TYPO3\CMS\Extbase\Utility\DebuggerUtility;
+
+use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
+use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithContentArgumentAndRenderStatic;
 
 /**
  * Class IsMandatoryFieldViewHelper
  *
  * @author Maximilian Fäßler <maximilian@faesslerweb.de>
- * @copyright Rkw Kompetenzzentrum
+ * @copyright RKW Kompetenzzentrum
  * @package RKW_RkwAuthors
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
+ * @todo write tests
  */
 class IsMandatoryFieldViewHelper extends \TYPO3\CMS\Fluid\Core\ViewHelper\AbstractViewHelper
 {
 
+    use CompileWithContentArgumentAndRenderStatic;
+
     /**
-     * return TRUE, if the given fieldName is NOT in given mandatoryFields (string-list from TypoScript)
-     * TRUE if optional
+     * Initialize arguments.
      *
-     * @param string $fieldName
-     * @param string $mandatoryFields
-     * @return boolean
+     * @return void
+     * @throws \TYPO3Fluid\Fluid\Core\ViewHelper\Exception
      */
-    public function render($fieldName, $mandatoryFields)
+    public function initializeArguments(): void
     {
+        parent::initializeArguments();
+        $this->registerArgument('fieldName', 'string', 'The current field name.', true);
+        $this->registerArgument('mandatoryFields', 'string', 'The comma-separated list of mandatory fields.', true);
+    }
+
+
+    /**
+     * Static rendering
+     *
+     * @param array $arguments
+     * @param \Closure $renderChildrenClosure
+     * @param RenderingContextInterface $renderingContext
+     * @return bool
+     */
+    static public function renderStatic(
+        array $arguments,
+        \Closure $renderChildrenClosure,
+        RenderingContextInterface $renderingContext
+    ): bool {
+
+        /** @var string $fieldName */
+        $fieldName = $arguments['fieldName'];
+
+        /** @var string $mandatoryFields */
+        $mandatoryFields = $arguments['mandatoryFields'];
+
         $mandatoryFieldsArray = array_map('trim', explode(',', $mandatoryFields));
         if (!in_array($fieldName, $mandatoryFieldsArray)) {
-
             return true;
-            //===
         }
 
         return false;
-        //===
     }
 
 
