@@ -27,12 +27,13 @@ $EM_CONF[$_EXTKEY] = [
     'modify_tables' => '',
     'clearCacheOnLoad' => 0,
     'lockType' => '',
-    'version' => '8.7.29',
+    'version' => '9.5.0',
 	'constraints' => [
 		'depends' => [
-            'typo3' => '8.7.0-8.7.99',
-            'rkw_basics' => '8.7.53-8.7.99',
-            'rkw_ajax' => '8.7.6-8.7.99',
+            'typo3' => '9.5.0-9.5.99',
+            'core_extended' => '9.5.4-9.5.99',
+            'accelerator' => '9.5.2-9.5.99',
+            'ajax_api' => '9.5.0-9.5.99',
 		],
 		'conflicts' => [
 		],
