@@ -9,22 +9,19 @@ call_user_func(
         //=================================================================
         \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
             'RKW.' . $extKey,
-            'Rkwauthors',
+            'main',
             array(
-                'Authors' => 'list, pageBox, pageBoxSmallFirst, pageBoxSmallRest, pageSchemaOrg, pageCommaList',
-                'Ajax' => 'filter'
+                'Authors' => 'list, pageBox, pageSchemaOrg, pageCommaList',
             ),
             // non-cacheable actions
             array(
                 'Authors' => '',
-
-                'Ajax' => ''
             )
         );
 
         \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
             'RKW.' . $extKey,
-            'Rkwauthorsdetail',
+            'details',
             array(
                 'Authors' => 'show, contactFormSend',
 
@@ -37,7 +34,7 @@ call_user_func(
 
         \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
             'RKW.' . $extKey,
-            'Rkwauthorsheadline',
+            'headline',
             array(
                 'Authors' => 'headline',
             ),
@@ -49,19 +46,7 @@ call_user_func(
 
         \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
             'RKW.' . $extKey,
-            'Rkwauthorswork',
-            array(
-                'Authors' => 'showWork',
-            ),
-            // non-cacheable actions
-            array(
-                'Authors' => 'showWork',
-            )
-        );
-
-        \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
-            'RKW.' . $extKey,
-            'Rkwauthorscontactbox',
+            'contact',
             array(
                 'Authors' => 'contactBox',
             ),
@@ -70,7 +55,6 @@ call_user_func(
                 'Authors' => 'contactBox',
             )
         );
-
 
         //=================================================================
         // Add Rootline Fields
